functions/requirements.txt
firebase-functions==0.1.1
firebase-admin==6.2.0
google-cloud-logging==3.9.0
deepgram-sdk==2.12.0 